Fix new libc++ compatibility
One problem was that libc_globals is no longer a POD.

Another one is about disallowed printing of char16_t to streams.

Bug: 175635923
Test: m MODULES-IN-bionic

+++ tests/wctype_test.cpp
@@ -36,7 +36,7 @@ static void TestIsWideFn(int fn(wint_t),
   UtfLocale l;
   for (const wchar_t* p = trues; *p; ++p) {
     if (!have_dl() && *p > 0x7f) {
-      GTEST_LOG_(INFO) << "skipping unicode test " << *p;
+      GTEST_LOG_(INFO) << "skipping unicode test " << static_cast<int>(*p);
       continue;
     }
     EXPECT_TRUE(fn(*p)) << *p;
@@ -44,7 +44,7 @@ static void TestIsWideFn(int fn(wint_t),
   }
   for (const wchar_t* p = falses; *p; ++p) {
     if (!have_dl() && *p > 0x7f) {
-      GTEST_LOG_(INFO) << "skipping unicode test " << *p;
+      GTEST_LOG_(INFO) << "skipping unicode test " << static_cast<int>(*p);
       continue;
     }
     EXPECT_FALSE(fn(*p)) << *p;
